requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

pydantic>=2.0.0
pydantic-settings>=2.0.0

passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

pyjwt>=2.8.0

python-dotenv>=1.0.0
email-validator>=2.0.0
python-jose[cryptography]
python-multipart

